refactor: Extract custom icon deletion logic (i906)
This commit extracts the logic for deleting custom icons into dedicated
use cases. The `resetGridItemCustomIcon` and
`resetEblanApplicationInfoCustomIcon` functions in their respective
repositories have been removed.

The new use cases, `DeleteGridItemCustomIconUseCase` and
`DeleteEblanApplicationInfoCustomIconUseCase`, encapsulate the file
deletion and repository update operations. This promotes code reuse
and simplifies the repository implementations.

Affected files:
- domain/repository/src/main/kotlin/com/eblan/launcher/domain/repository/EblanApplicationInfoRepository.kt
- domain/repository/src/main/kotlin/com/eblan/launcher/domain/repository/GridRepository.kt
- data/repository/src/main/kotlin/com/eblan/launcher/data/repository/DefaultEblanApplicationInfoRepository.kt
- data/repository/src/main/kotlin/com/eblan/launcher/data/repository/DefaultGridRepository.kt
- domain/use-case/src/main/kotlin/com/eblan/launcher/domain/usecase/grid/DeleteGridItemCustomIconUseCase.kt
- domain/use-case/src/main/kotlin/com/eblan/launcher/domain/usecase/application/DeleteEblanApplicationInfoCustomIconUseCase.kt
- feature/edit-grid-item/src/main/kotlin/com/eblan/launcher/feature/editgriditem/EditGridItemViewModel.kt
- feature/edit-application-info/src/main/kotlin/com/eblan/launcher/feature/editapplicationinfo/EditApplicationInfoViewModel.kt
