+++ domain/repository/src/main/kotlin/com/eblan/launcher/domain/repository/EblanApplicationInfoRepository.kt
@@ -45,8 +45,6 @@ interface EblanApplicationInfoRepository {
 
     suspend fun updateEblanApplicationInfo(eblanApplicationInfo: EblanApplicationInfo)
 
-    suspend fun resetEblanApplicationInfoCustomIcon(eblanApplicationInfo: EblanApplicationInfo)
-
     suspend fun getEblanApplicationInfoByComponentName(
         serialNumber: Long,
         componentName: String,
