+++ feature/edit-grid-item/src/main/kotlin/com/eblan/launcher/feature/editgriditem/EditGridItemViewModel.kt
@@ -30,6 +30,7 @@ import com.eblan.launcher.domain.model.IconPackInfoComponent
 import com.eblan.launcher.domain.model.PackageManagerIconPackInfo
 import com.eblan.launcher.domain.repository.GridRepository
 import com.eblan.launcher.domain.usecase.application.GetEblanApplicationInfosUseCase
+import com.eblan.launcher.domain.usecase.grid.DeleteGridItemCustomIconUseCase
 import com.eblan.launcher.domain.usecase.grid.GetGridItemsUseCase
 import com.eblan.launcher.domain.usecase.grid.UpdateGridItemCustomIconUseCase
 import com.eblan.launcher.feature.editgriditem.model.EditGridItemUiState
@@ -55,6 +56,7 @@ internal class EditGridItemViewModel @Inject constructor(
     getEblanApplicationInfosUseCase: GetEblanApplicationInfosUseCase,
     private val getGridItemsUseCase: GetGridItemsUseCase,
     private val updateGridItemCustomIconUseCase: UpdateGridItemCustomIconUseCase,
+    private val deleteGridItemCustomIconUseCase: DeleteGridItemCustomIconUseCase,
     @param:Dispatcher(EblanDispatchers.Default) private val defaultDispatcher: CoroutineDispatcher,
 ) : ViewModel() {
     private val editGridItemRouteData = savedStateHandle.toRoute<EditGridItemRouteData>()
@@ -107,7 +109,7 @@ internal class EditGridItemViewModel @Inject constructor(
 
     fun resetGridItemCustomIcon(gridItem: GridItem) {
         viewModelScope.launch {
-            gridRepository.resetGridItemCustomIcon(gridItem = gridItem)
+            deleteGridItemCustomIconUseCase(gridItem = gridItem)
 
             getGridItem()
         }
